some changes for image update functionality

+++ src/main/java/com/epam/edumanagementsystem/student/rest/service/impl/StudentServiceImpl.java
@@ -8,7 +8,6 @@ import com.epam.edumanagementsystem.student.model.entity.Student;
 import com.epam.edumanagementsystem.student.rest.api.StudentController;
 import com.epam.edumanagementsystem.student.rest.repository.StudentRepository;
 import com.epam.edumanagementsystem.student.rest.service.StudentService;
-import com.epam.edumanagementsystem.teacher.model.dto.TeacherDto;
 import com.epam.edumanagementsystem.util.entity.User;
 import com.epam.edumanagementsystem.util.exceptions.UserNotFoundException;
 import com.epam.edumanagementsystem.util.imageUtil.rest.service.ImageService;
@@ -133,15 +132,12 @@ public class StudentServiceImpl implements StudentService {
 
     @Override
     public void addProfilePicture(StudentDto studentDto, MultipartFile multipartFile) {
+        if (studentDto.getPicUrl() != null) {
+            imageService.deleteImage(studentDto.getPicUrl());
+        }
         studentDto.setPicUrl(imageService.saveImage(multipartFile));
         logger.debug(studentDto.getPicUrl());
         updateForPic(studentDto);
-    }
-
-    @Override
-    public void updateImage(StudentDto studentDto, MultipartFile multipartFile) {
-        imageService.deleteImage(studentDto.getPicUrl());
-        addProfilePicture(studentDto,multipartFile);
     }
 
     @Override
